requirements.txt
django
django-rest-framework
gunicorn
djangorestframework-simplejwt
uuid